Color Diffusion/Color-diffusion/requirements.txt
accelerate
accelerate==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
denoising-diffusion-pytorch
docker-pycreds==0.4.0
einops
einops_exts
einops==0.8.1
ema-pytorch
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gdown
gitdb==4.0.12
GitPython==3.1.44
helpers
huggingface-hub==0.28.1
icecream
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
kiwisolver==1.4.8
labml_nn
layers
lazy_loader==0.4
lightning-utilities==0.12.0
MarkupSafe==3.0.2
matplotlib
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy
numpy==2.2.2
omegaconf
omegaconf==2.3.0
packaging==24.2
pandas==2.2.3
Pillow
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pydantic_core==2.27.2
pydantic==2.10.6
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch_lightning
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.2
scikit_image
scikit-image==0.25.1
scipy==1.15.1
sentry-sdk==2.20.0
setproctitle==1.3.4
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tensorboard
tifffile==2025.1.10
torch
torch==2.6.0
torch-ema
torch-ema==0.3
torchinfo
torchinfo==1.8.0
torchmetrics==1.6.1
torchvision
torchvision==0.21.0
tqdm
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wandb
wandb==0.19.6
yarl==1.18.3
